add a spinner when is uploading image and extract Preview in separate file
